More remote control stuff, desktop side is almost complete (possibly is) Neatened code, made adding of the comments more robust Fixed deleting latest added track issue

+++ Utility/Remote/RemoteInterprocessConnection.h
@@ -31,8 +31,8 @@ public:
     void sendPlayingData();
     void sendAlbumArt();
     void sendLength (double length);
-    void sendPosition (int position);
-    void sendVolume (float volume);
+    void sendPosition (double position);
+    void sendVolume (double volume);
     void sendPlayState();
 
 private:
